# reference/apisperu-swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "API REST Facturaci\u00f3n Electr\u00f3nica - APIsPERU. <br /> Si se siente m\u00e1s c\u00f3modo usando Swagger-UI, puede <a target='_blank' href='/doc/swagger/ui'>ingresar aqu\u00ed.</a>",
        "version": "1.3",
        "title": "API DE FACTURACI\u00d3N",
        "contact": {
            "name": "Apisperu",
            "email": "soporte@apisperu.com"
        },
        "x-logo": {
            "url": "/img/logoredoc.png",
            "altText": "Logo APIsPERU"
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ],
    "paths": {
...
        "/summary/send": {
            "post": {
                "tags": [
                    "summary"
                ],
                "summary": "Enviar una Resumen diario de Boletas",
                "description": "Retorna la respuesta de SUNAT y el XML firmado",
                "operationId": "sendSummary",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Summary"
                },
                "responses": {
                    "200": {
                        "description": "Operacion exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SummaryResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de Validaci\u00f3n",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ValidationResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/summary/xml": {
            "post": {
                "tags": [
                    "summary"
                ],
                "summary": "Generar el XML de un resumen diario",
                "description": "Retorna el XML",
                "operationId": "getSummaryXml",
                "requestBody": {
                    "$ref": "#/components/requestBodies/Summary"
                },
                "responses": {
                    "200": {
                        "description": "Operacion exitosa"
                    },
                    "400": {
                        "description": "Error de Validaci\u00f3n",
                        "content": {
                            "text/xml": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ValidationResponse"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/summary/status": {
            "get": {
                "tags": [
                    "summary"
                ],
                "summary": "Obtiene el estado del ticket",
                "description": "Retorna el XML",
                "operationId": "getSummaryStatus",
                "parameters": [
                    {
                        "name": "ticket",
                        "in": "query",
                        "description": "Ticket de Resumen diario",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ruc",
                        "in": "query",
                        "description": "RUC de empresa (Multiples Empresas)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Operacion exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StatusResult"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "https://facturacion.apisperu.com/api/v1"
        }
    ],
    "components": {
        "requestBodies": {
...
            "Summary": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/Summary"
                        },
                        "examples": {
                            "example1": {
                                "$ref": "#/components/examples/resumen"
                            },
                            "example2": {
                                "$ref": "#/components/examples/resumenContingencia"
                            },
                            "example3": {
                                "$ref": "#/components/examples/resumenUSD"
                            }
                        }
                    }
                },
                "description": "Comprobante",
                "required": true
            },
...
        },
        "examples": {
...
            "resumenContingencia": {
                "summary": "Resumen diario contingencia",
                "value": {
                    "fecGeneracion": "2021-01-27T00:00:00-05:00",
                    "fecResumen": "2021-01-27T00:00:00-05:00",
                    "correlativo": "001",
                    "moneda": "PEN",
                    "company": {
                        "ruc": 20000000001,
                        "razonSocial": "Mi empresa",
                        "nombreComercial": "Mi empresa",
                        "address": {
                            "direccion": "Direccion empresa",
                            "provincia": "LIMA",
                            "departamento": "LIMA",
                            "distrito": "LIMA",
                            "ubigueo": "150101"
                        }
                    },
                    "details": [
                        {
                            "tipoDoc": "03",
                            "serieNro": "0001-1",
                            "estado": "3",
                            "clienteTipo": "1",
                            "clienteNro": "00000000",
                            "total": 119.555,
                            "mtoOperGravadas": 20,
                            "mtoOperInafectas": 24.4,
                            "mtoOperExoneradas": 50,
                            "mtoOtrosCargos": 21.555,
                            "mtoIGV": 3.6
                        }
                    ]
                }
            },
            "resumenUSD": {
                "summary": "Resumen diario USD",
                "value": {
                    "fecGeneracion": "2021-01-27T00:00:00-05:00",
                    "fecResumen": "2021-01-27T00:00:00-05:00",
                    "correlativo": "211",
                    "moneda": "USD",
                    "company": {
                        "ruc": 20000000001,
                        "razonSocial": "Mi empresa",
                        "nombreComercial": "Mi empresa",
                        "address": {
                            "direccion": "Direccion empresa",
                            "provincia": "LIMA",
                            "departamento": "LIMA",
                            "distrito": "LIMA",
                            "ubigueo": "150101"
                        }
                    },
                    "details": [
                        {
                            "tipoDoc": "03",
                            "serieNro": "B001-1",
                            "estado": "3",
                            "clienteTipo": "1",
                            "clienteNro": "00000000",
                            "total": 129,
                            "mtoOperGravadas": 20,
                            "mtoOperInafectas": 24.4,
                            "mtoOperExoneradas": 50,
                            "mtoOperExportacion": 10,
                            "mtoOtrosCargos": 21,
                            "mtoIGV": 3.6
                        }
                    ]
                }
            },
            "resumen": {
                "summary": "Resumen diario",
                "value": {
                    "fecGeneracion": "2021-01-27T00:00:00-05:00",
                    "fecResumen": "2021-01-27T00:00:00-05:00",
                    "correlativo": "001",
                    "moneda": "PEN",
                    "company": {
                        "ruc": 20000000001,
                        "razonSocial": "Mi empresa",
                        "nombreComercial": "Mi empresa",
                        "address": {
                            "direccion": "Direccion empresa",
                            "provincia": "LIMA",
                            "departamento": "LIMA",
                            "distrito": "LIMA",
                            "ubigueo": "150101"
                        }
                    },
                    "details": [
                        {
                            "tipoDoc": "03",
                            "serieNro": "B001-1",
                            "estado": "3",
                            "clienteTipo": "1",
                            "clienteNro": "00000000",
                            "total": 129.555,
                            "mtoOperGravadas": 20,
                            "mtoOperInafectas": 24.4,
                            "mtoOperExoneradas": 50,
                            "mtoOperExportacion": 10.555,
                            "mtoOtrosCargos": 21,
                            "mtoIGV": 3.6
                        },
                        {
                            "tipoDoc": "07",
                            "serieNro": "B001-4",
                            "docReferencia": {
                                "tipoDoc": "03",
                                "nroDoc": "0001-122"
                            },
                            "estado": "1",
                            "clienteTipo": "1",
                            "clienteNro": "00000000",
                            "total": 200,
                            "mtoOperGravadas": 40,
                            "mtoOperExoneradas": 30,
                            "mtoOperInafectas": 120,
                            "mtoIGV": 7.2,
                            "mtoISC": 2.8
                        },
                        {
                            "tipoDoc": "03",
                            "serieNro": "B001-2",
                            "estado": "1",
                            "clienteTipo": "1",
                            "clienteNro": "00000000",
                            "percepcion": {
                                "codReg": "01",
                                "tasa": 2,
                                "mtoBase": 119,
                                "mto": 2,
                                "mtoTotal": 121.38
                            },
                            "total": 119,
                            "mtoOperGravadas": 20,
                            "mtoOperInafectas": 24.4,
                            "mtoOperExoneradas": 50,
                            "mtoOtrosCargos": 21,
                            "mtoIGV": 3.6
                        }
                    ]
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Value: Bearer {jwt}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "Address": {
                "type": "object",
                "properties": {
                    "ubigueo": {
                        "type": "string"
                    },
                    "codigoPais": {
                        "type": "string"
                    },
                    "departamento": {
                        "type": "string"
                    },
                    "provincia": {
                        "type": "string"
                    },
                    "distrito": {
                        "type": "string"
                    },
                    "urbanizacion": {
                        "type": "string"
                    },
                    "direccion": {
                        "type": "string"
                    },
                    "codLocal": {
                        "type": "string"
                    }
                }
            },
...
            "Company": {
                "type": "object",
                "properties": {
                    "ruc": {
                        "type": "string"
                    },
                    "razonSocial": {
                        "type": "string"
                    },
                    "nombreComercial": {
                        "type": "string"
                    },
                    "address": {
                        "$ref": "#/components/schemas/Address"
                    },
                    "email": {
                        "type": "string"
                    },
                    "telephone": {
                        "type": "string"
                    }
                }
            },
...
            "Document": {
                "type": "object",
                "properties": {
                    "tipoDoc": {
                        "description": "<a href='http://www.sunat.gob.pe/legislacion/superin/2017/anexoVII-117-2017.pdf' target='_blank'>Cat\u00e1logo No.  21: Documentos Relacionados</a>",
                        "type": "string"
                    },
                    "nroDoc": {
                        "type": "string"
                    }
                }
            },
...
            "SummaryPerception": {
                "type": "object",
                "properties": {
                    "codReg": {
                        "type": "string"
                    },
                    "tasa": {
                        "type": "number",
                        "format": "float"
                    },
                    "mtoBase": {
                        "type": "number",
                        "format": "float"
                    },
                    "mto": {
                        "type": "number",
                        "format": "float"
                    },
                    "mtoTotal": {
                        "type": "number",
                        "format": "float"
                    }
                }
            },
            "SummaryDetail": {
                "type": "object",
                "properties": {
                    "tipoDoc": {
                        "description": "<a href='http://www.sunat.gob.pe/legislacion/superin/2017/anexoVII-117-2017.pdf' target='_blank'>Cat\u00e1logo No. 01</a>",
                        "type": "string"
                    },
                    "serieNro": {
                        "type": "string"
                    },
                    "clienteTipo": {
                        "type": "string"
                    },
                    "clienteNro": {
                        "type": "string"
                    },
                    "docReferencia": {
                        "$ref": "#/components/schemas/Document"
                    },
                    "percepcion": {
                        "$ref": "#/components/schemas/SummaryPerception"
                    },
                    "estado": {
                        "description": "<a href='http://www.sunat.gob.pe/legislacion/superin/2019/anexoV-114-2019.pdf' target='_blank'>No. 04</a>",
                        "type": "string"
                    },
                    "total": {
                        "type": "number",
                        "format": "float"
                    },
                    "porcentajeIgv": {
                        "type": "number",
                        "format": "float"
                    },
                    "mtoOperGravadas": {
                        "type": "number",
                        "format": "float"
                    },
                    "mtoOperInafectas": {
                        "type": "number",
                        "format": "float"
                    },
                    "mtoOperExoneradas": {
                        "type": "number",
                        "format": "float"
                    },
                    "mtoOperExportacion": {
                        "type": "number",
                        "format": "float"
                    },
                    "mtoOperGratuitas": {
                        "type": "number",
                        "format": "float"
                    },
                    "mtoOtrosCargos": {
                        "type": "number",
                        "format": "float"
                    },
                    "mtoIGV": {
                        "type": "number",
                        "format": "float"
                    },
                    "mtoIvap": {
                        "type": "number",
                        "format": "float"
                    },
                    "mtoIcbper": {
                        "type": "number",
                        "format": "float"
                    },
                    "mtoISC": {
                        "type": "number",
                        "format": "float"
                    },
                    "mtoOtrosTributos": {
                        "type": "number",
                        "format": "float"
                    }
                }
            },
            "Summary": {
                "type": "object",
                "properties": {
                    "correlativo": {
                        "type": "string"
                    },
                    "fecGeneracion": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "fecResumen": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "moneda": {
                        "type": "string"
                    },
                    "company": {
                        "$ref": "#/components/schemas/Company"
                    },
                    "details": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SummaryDetail"
                        }
                    },
                    "xmlId": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    }
                }
            },
...
            "Error": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
            "CdrResponse": {
                "type": "object",
                "properties": {
                    "accepted": {
                        "type": "boolean",
                        "readOnly": true
                    },
                    "id": {
                        "type": "string"
                    },
                    "code": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "notes": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
...
            "SummaryResult": {
                "type": "object",
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "error": {
                        "$ref": "#/components/schemas/Error"
                    },
                    "ticket": {
                        "type": "string"
                    }
                }
            },
            "StatusResult": {
                "type": "object",
                "properties": {
                    "cdrZip": {
                        "type": "string",
                        "format": "byte"
                    },
                    "cdrResponse": {
                        "$ref": "#/components/schemas/CdrResponse"
                    },
                    "success": {
                        "type": "boolean"
                    },
                    "error": {
                        "$ref": "#/components/schemas/Error"
                    },
                    "code": {
                        "type": "string"
                    }
                }
            },
...
            "SummaryResponse": {
                "type": "object",
                "properties": {
                    "xml": {
                        "type": "string"
                    },
                    "hash": {
                        "type": "string",
                        "description": "Resumen de la Firma Digital"
                    },
                    "sunatResponse": {
                        "$ref": "#/components/schemas/SummaryResult"
                    }
                }
            },
            "ValidationResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "field": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}